# doc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Keymaster API",
    "version": "1.0.0",
    "description": "API for managing wallets, identities, credentials, and more"
  },
...
  "paths": {
...
    "/registries": {
      "get": {
        "summary": "List all registries",
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "example": [
                  "local",
                  "hyperswarm",
                  "TESS"
                ]
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet": {
      "get": {
        "summary": "Load wallet",
        "tags": [
          "Wallet"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "seed": {
                      "type": "object",
                      "properties": {
                        "mnemonic": {
                          "type": "string",
                          "example": "string"
                        },
                        "hdkey": {
                          "type": "object",
                          "properties": {
                            "xpriv": {
                              "type": "string",
                              "example": "string"
                            },
                            "xpub": {
                              "type": "string",
                              "example": "string"
                            }
                          }
                        }
                      }
                    },
                    "counter": {
                      "type": "integer",
                      "example": 2
                    },
                    "ids": {
                      "type": "object",
                      "properties": {
                        "example-id": {
                          "type": "object",
                          "properties": {
                            "did": {
                              "type": "string",
                              "example": "did:test:string"
                            },
                            "account": {
                              "type": "integer",
                              "example": 1
                            },
                            "index": {
                              "type": "integer",
                              "example": 0
                            }
                          }
                        }
                      }
                    },
                    "names": {
                      "type": "object",
                      "properties": {
                        "example-name-1": {
                          "type": "string",
                          "example": "did:string"
                        },
                        "example-name-2": {
                          "type": "string",
                          "example": "did:string"
                        },
                        "example-name-3": {
                          "type": "string",
                          "example": "did:string"
                        }
                      }
                    },
                    "current": {
                      "type": "string",
                      "example": "example-id"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Save wallet",
        "tags": [
          "Wallet"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "wallet": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet/new": {
      "post": {
        "summary": "Create a new wallet",
        "tags": [
          "Wallet"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mnemonic": {
                    "type": "string"
                  },
                  "overwrite": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet/backup": {
      "post": {
        "summary": "Backup wallet",
        "tags": [
          "Wallet"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet/recover": {
      "post": {
        "summary": "Recover wallet",
        "tags": [
          "Wallet"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet/check": {
      "post": {
        "summary": "Check wallet",
        "tags": [
          "Wallet"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet/fix": {
      "post": {
        "summary": "Fix wallet",
        "tags": [
          "Wallet"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet/mnemonic": {
      "get": {
        "summary": "Get decrypted mnemonic",
        "tags": [
          "Wallet"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "equip vital solution badge crucial pitch excite fork favorite soldier scene tray"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/ids/current": {
      "get": {
        "summary": "Get current ID",
        "tags": [
          "Identity"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Set current ID",
        "tags": [
          "Identity"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/ids": {
      "get": {
        "summary": "List all IDs",
        "tags": [
          "Identity"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/ids/new": {
      "post": {
        "summary": "Create a new ID",
        "tags": [
          "Identity"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "registry": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/ids/{id}/backup": {
      "post": {
        "summary": "Backup an ID",
        "tags": [
          "Identity"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/ids/{did}/recover": {
      "post": {
        "summary": "Recover an DID",
        "tags": [
          "Identity"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "did",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "did": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/names": {
      "get": {
        "summary": "List all names",
        "tags": [
          "Name"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Add a new name",
        "tags": [
          "Name"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "did": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/credentials/issued": {
      "get": {
        "summary": "List all issued credentials",
        "tags": [
          "Credential"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
  }
}